+++ packages/retry-config-provider/src/index.ts
@@ -1,1 +1,1 @@
-export * from "./maxAttemptsProvider";
+export * from "./defaultProvider";
